requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
dataclasses-json==0.5.14
datasets==2.10.1
dill==0.3.6
dnspython==2.4.2
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.9.2
greenlet==2.0.2
huggingface-hub==0.17.3
idna==3.4
langchain==0.0.292
langsmith==0.0.41
loguru==0.7.2
marshmallow==3.20.1
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
numexpr==2.8.7
numpy==1.26.0
openai==0.28.0
packaging==23.1
pandas==2.1.1
pinecone-client==2.2.4
pyarrow==13.0.0
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
responses==0.18.0
six==1.16.0
SQLAlchemy==2.0.21
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
xxhash==3.3.0
yarl==1.9.2